Update random dog command

+++ src/commands/Random/rng.js
@@ -32,6 +32,7 @@ exports.run = async (client, message, args, level) => {
 		default:
 		case 'easy':
 			settings = {
+				mode: 1,
 				max: 10,
 				time: 10
 			};
@@ -40,14 +41,24 @@ exports.run = async (client, message, args, level) => {
 		case 'med':
 		case 'mid':
 			settings = {
+				mode: 1,
 				max: 50,
 				time: 30
 			};
 			break;
 		case 'hard':
 			settings = {
+				mode: 1,
+				max: 75,
+				time: 45
+			};
+			break;
+		case 'impossible':
+		case 'i':
+			settings = {
+				mode: 1,
 				max: 100,
-				time: 30
+				time: 45
 			};
 			break;
 	}
@@ -59,26 +70,31 @@ exports.run = async (client, message, args, level) => {
 		.setTitle(`RNG (${difficulty})`)
 		.setColor(client.colors.default)
 		.setDescription(`I have chosen a number between **1** and **${settings.max}**\n\nEveryone in this channel can guess by typing a number.\n\nThe game will end if __${time} seconds__ has elapsed.\n\n**GO**.`);
-	
+
 	const msg = await message.channel.send({ embed });
-	message.channel.awaitMessages(m => m.cleanContent.trim() === `${chosenNumber}`, { max: 1, time: (time*1000), errors: ['time'] }).then(col => {
-		const winner = col.first().member;
-		embed = new RichEmbed()
-			.setTitle('RNG')
-			.setColor(client.colors.green)
-			.setDescription(`<@${winner.user.id}> has guessed the number first (0 - ${settings.max})!\n**The number was __${chosenNumber}__!**`);
-		msg.delete().then(m => {
-			m.channel.send({ embed })
+
+	if (settings.mode < 2) {
+		message.channel.awaitMessages(m => m.cleanContent.trim() === `${chosenNumber}`, { max: 1, time: (time*1000), errors: ['time'] }).then(col => {
+			const winner = col.first().member;
+			embed = new RichEmbed()
+				.setTitle('RNG')
+				.setColor(client.colors.green)
+				.setDescription(`<@${winner.user.id}> has guessed the number first (0 - ${settings.max})!\n**The number was __${chosenNumber}__!**`);
+			msg.delete().then(m => {
+				m.channel.send({ embed })
+			});
+		}).catch(() => {
+			embed = new RichEmbed()
+				.setTitle('RNG')
+				.setColor(client.colors.red)
+				.setDescription(`No one guessed the number (0 - ${settings.max})!\n**The number was __${chosenNumber}__!**`);
+			msg.delete().then(m => {
+				m.channel.send({ embed })
+			});
 		});
-	}).catch(() => {
-		embed = new RichEmbed()
-			.setTitle('RNG')
-			.setColor(client.colors.red)
-			.setDescription(`No one guessed the number (0 - ${settings.max})!\n**The number was __${chosenNumber}__!**`);
-		msg.delete().then(m => {
-			m.channel.send({ embed })
-		});
-	});
+	} else {
+
+	}
 };
 
 exports.conf = {
